+++ package.json
@@ -48,6 +48,7 @@
     "react-dom": "*"
   },
   "devDependencies": {
+    "@types/jest": "^25.2.1",
     "@types/react": "^16.8.19",
     "@types/react-dom": "^16.8.4",
     "@types/warning": "^3.0.0",
